+++ mmodel/modifier.py
@@ -1,6 +1,7 @@
 from functools import wraps
 from inspect import signature, Parameter, Signature
 from types import MappingProxyType
+from string import Formatter
 
 
 def format_parameters(*args, **kwargs) -> list:
@@ -14,7 +15,7 @@ def format_parameters(*args, **kwargs) -> list:
     return param_str
 
 
-def modifier(name, *args, **kwargs):
+def add_modifier_metadata(name, *args, **kwargs):
     """Decorator to add metadata to a function."""
 
     def decorator(func):
@@ -36,7 +37,7 @@ def loop_input(parameter: str):
         The target parameter name is changed to f"{param}_loop"
     """
 
-    @modifier("loop_input", parameter=parameter)
+    @add_modifier_metadata("loop_input", parameter=parameter)
     def loop(func):
         param_list = []
         for param in signature(func).parameters.values():
@@ -68,7 +69,7 @@ def zip_loop_inputs(parameters: list):
         provided, the parameters should be delimited by ", ".
     """
 
-    @modifier("zip_loop_inputs", parameters=parameters)
+    @add_modifier_metadata("zip_loop_inputs", parameters=parameters)
     def zip_loop(func):
         @wraps(func)
         def loop_wrapped(**kwargs):
@@ -107,7 +108,7 @@ def profile_time(number=1, repeat=1, verbose=False, precision=2):
 
     timer = timeit.default_timer
 
-    @modifier(
+    @add_modifier_metadata(
         "zip_loop_inputs",
         number=number,
         repeat=repeat,
@@ -140,3 +141,87 @@ def profile_time(number=1, repeat=1, verbose=False, precision=2):
         return wrapped
 
     return timeit_modifier
+
+
+def parse_fields(format_str):
+    """Parse the field from the format string.
+
+    :param str format_str: format string
+    :return: list of fields
+
+    The function parses out the field names in the format string.
+    Some field names have slicers or attribute access, such as
+    B0.value, B0[0], B0[0:2]. The function only returns B0 for all
+    these fields. Since there can be duplicated fields after the
+    name split, the function returns unique elements.
+    """
+
+    # this is an internal function for Formatter
+    # consider rewriting with custom function to prevent breaking
+    # the function ignores slicing and attribute access
+    # B0.value -> B0, B0[0] -> B0
+    from _string import formatter_field_name_split
+
+    fields = [
+        formatter_field_name_split(field)[0]
+        for _, field, _, _ in Formatter().parse(format_str)
+        if field
+    ]
+    return list(set(fields))  # return unique elements
+
+
+def print_inputs(format_str: str, **pargs):
+    """Print the node input to the console.
+
+    :param str stdout_format: format string for input and output
+        The format should be keyword only.
+    :param pargs: keyword arguments for the print function
+
+    The names of the parameters are parsed from the format string.
+    """
+
+    @add_modifier_metadata("print_inputs", format_str=format_str, **pargs)
+    def stdout_inputs_modifier(func):
+        inputs = parse_fields(format_str)
+
+        @wraps(func)
+        def wrapped(**kwargs):
+            """Print input parameter."""
+            input_dict = {k: kwargs[k] for k in inputs}
+            print(format_str.format(**input_dict), **pargs)
+            return func(**kwargs)
+
+        return wrapped
+
+    return stdout_inputs_modifier
+
+
+def print_output(format_str: str, **pargs):
+    """Print the node output to the console.
+
+    :param str stdout_format: format string for input and output
+        The format should be keyword only. The behavior is for keeping the
+        consistency with other print modifiers.
+    :param str end: end of printout
+
+    The names of the parameters are parsed from the format string. The
+    use of the stdout_format is different from the input method, as
+    the modifiers do not know the return name of the node. Only one
+    output field is allowed and the field name is used as the return name.
+    """
+
+    @add_modifier_metadata("print_output", format_str=format_str, **pargs)
+    def stdout_output_modifier(func):
+        output = parse_fields(format_str)[0]
+
+        @wraps(func)
+        def wrapped(**kwargs):
+            """Print output parameter."""
+
+            result = func(**kwargs)
+            print(format_str.format(**{output: result}), **pargs)
+            return result
+
+        return wrapped
+
+    return stdout_output_modifier